+++ src/app.ts
@@ -1,6 +1,7 @@
 import 'dotenv/config'
 import Bot from './bot'
 import { loadAccounts } from './fava'
+import { loadShortcuts } from './shortcuts'
 
 export let ACCOUNTS: string[] = []
 
@@ -11,6 +12,7 @@ start().catch(err => {
 
 async function start () {
   ACCOUNTS = await loadAccounts()
+  await loadShortcuts()
   const allowedIds = process.env.ALLOWED_USER_IDS!.split(',').map(Number).filter(e => !isNaN(e))
   const bot = new Bot(process.env.TELEGRAM_TOKEN!, { allowedIds })
   bot.start()