+++ src/bot.ts
@@ -1,10 +1,11 @@
 import TelegramBot from 'node-telegram-bot-api'
-import { CANCEL, GET_ERRORS, NEW_BALANCE, NEW_NOTE, NEW_TRANSACTION } from './consts'
+import { CANCEL, GET_ERRORS, NEW_BALANCE, NEW_NOTE, NEW_TRANSACTION, USE_SHORTCUT } from './consts'
 import { DEFAULT_KEYBOARD, NO_KEYBOARD } from './markup'
 import newTransaction from './state_machines/newTransaction'
 import newBalance from './state_machines/newBalance'
 import newNote from './state_machines/newNote'
 import getErrors from './state_machines/getErrors'
+import useShortcut from './state_machines/useShortcut'
 
 type BotOptions = {
   allowedIds: number[]
@@ -53,6 +54,9 @@ export default class Bot {
             break
           case GET_ERRORS:
             this._machines[msg.chat.id] = getErrors(msg, this._client)
+            break
+          case USE_SHORTCUT:
+            this._machines[msg.chat.id] = useShortcut(msg, this._client)
             break
           default:
             return this._client.sendMessage(msg.chat.id, '👋 Hi there!', DEFAULT_KEYBOARD)