+++ src/markup.ts
@@ -44,12 +44,12 @@ export const CANCEL_OR_DONE_KEYBOARD = {
 
 export const NO_KEYBOARD = { reply_markup: { remove_keyboard: true } }
 
-export const accountsKeyboard = (done: boolean) => ({
+export const accountsKeyboard = (done: boolean, filter?: string) => ({
   reply_markup: {
     resize_keyboard: true,
     keyboard: [
       done ? [{ text: DONE }, { text: CANCEL }] : [{ text: CANCEL }],
-      ...(ACCOUNTS.map(e => [{ text: e }]))
+      ...(ACCOUNTS.filter(e => e.startsWith(filter || '')).map(e => [{ text: e }]))
     ]
   }
 })