+++ src/app.ts
@@ -1,6 +1,6 @@
 import 'dotenv/config'
 import Bot from './bot'
-import { loadAccounts } from './fava'
+import { loadLedgerData } from './fava'
 import { loadShortcuts } from './shortcuts'
 
 export let ACCOUNTS: string[] = []
@@ -11,9 +11,14 @@ start().catch(err => {
 })
 
 async function start () {
-  ACCOUNTS = await loadAccounts()
+  const ledgerData = await loadLedgerData()
+  ACCOUNTS = ledgerData.accounts
+  process.env.DEFAULT_CURRENCY = process.env.DEFAULT_CURRENCY || ledgerData.operatingCurrency[0] || 'EUR'
+
   await loadShortcuts()
+
   const allowedIds = process.env.ALLOWED_USER_IDS!.split(',').map(Number).filter(e => !isNaN(e))
   const bot = new Bot(process.env.TELEGRAM_TOKEN!, { allowedIds })
+
   bot.start()
 }